Fix analysis/viewer inconsistency in rprec results
Previously, instructions that ended up failing all reduced-precision tests
(i.e., all 52 bits are required) were given a tag of "IGNORE". This makes the
final configuration more efficient since IGNORE and RPREC-52 instrumentation are
functionally equivalent but the former doesn't actually insert any code.
However, this was causing a problem in the viewer because it is largely
mode-agnostic and there's no clean way to hard-code a precision level of 52 bits
for instructions with an "IGNORE" tag if the config file was an RPREC file. The
cleaner fix is to change the analysis to just emit an RPREC-52 tag for those
instructions. It's less efficient for the final configuration test but provides
a more consistent view in the GUI. This should also resolve the "zero precision"
anomalies in the execution histogram output for some applications.