Keep major data structures in memory instead of files
This is another major change in how configurations are run. Before the
recent testing routine redesign (31c5978ca8987e8b712dbd46ce36fa3ef9c050ac),
all of the worker processes shared the primary data structures (work
queue, in-process tracker, and tested configuration archive), using file
locking for synchronization.

This did not not scale well once the search must run thousands of
configurations, and now that there is only a single monitor process
there is no need for interprocess synchronization. Accordingly, the
major data structures are now stored in memory and only written to disk
after a change. This significantly improves access times and reduces the
amount of I/O traffic incurred by a search.

As a side effect, this also simplifies the resuming code and makes it
more robust. I have also taken this opportunity to remove the old
caching code because the tested configuration archive is essentially a
cache. With more reliable resumption there is no longer a compelling
case for a separate "prior results" cache.