Change config launch/wait routines
While testing with an Ubuntu AWS image, I observed that the "exec" command was
not working the way I expected (and the way it had been working on all my other
test machines). The output redirect wasn't working, and it was causing the
spawned process to go defunct while a child process continued running. This
broke the wait-for-process-to-finish routine. I do not know for sure if this is
an artifact of Ubuntu specifically or if it's just a difference in versions of
Ruby, but the new implementation seems to work in all cases.