Fix bug in ddebug strategy
In lines 5-6 of Figure 4 in [Rubio-Gonzalez2013SC], new changesets
are formed based on spliting Delta and then using it to calculate
a complement for each split. Our implementation used @lc (our
equivalent to Delta) for the spitting, but then reverted to
@all_configs when calculating the complement. This will produce
correct results at the beginning of the search (when Delta =
@all_configs) but will fail once we find a new LC (lower cost)
changeset.

This bug did not affect the correctness of any results in the sense
that any valid configurations found previously are still valid and any
invalid configurations are still invalid. However, the search may not
have been finding as many speedups because it might continue to keep
some variables in higher precision when it has already found them
to be replaceable.

This strategy was already an effective algorithm, but it should
find even better results now.